Fixed types.

+++ mask2former/maskformer_model_ml.py
@@ -410,11 +410,11 @@ class MaskFormerML(nn.Module):
         print("Metaloss tar labels shape: {}".format(tar[0]['labels'].shape))
         all_targets = []
         for i in range(len(tar)):
-            t_mask = tar[i]['masks']
-            t_label = tar[i]['labels']
+            t_mask = tar[i]['masks'].float()
+            t_label = tar[i]['labels'].float()
             t_mask_label = torch.einsum("qhw,q->hw", t_mask, t_label)
             all_targets.append(t_mask_label)
-        batched_target = torch.stack(all_targets, dim=0)
+        batched_target = torch.stack(all_targets, dim=0).long()
         print("Metaloss b-target shape: {}".format(batched_target.shape))
         print("Metaloss b-target max: {}".format(batched_target.max()))
         mask_cls_results = out["pred_logits"].detach()
